+++ src/components/layout/MainWindow.tsx
@@ -19,6 +19,7 @@ import { SidebarWidthProvider } from './SidebarWidthContext'
 import { MainWindowContent } from './MainWindowContent'
 import { CommandPalette } from '@/components/command-palette/CommandPalette'
 import { BranchConflictDialog } from '@/components/worktree/BranchConflictDialog'
+import { ClearContextConfirmDialog } from '@/components/chat/ClearContextConfirmDialog'
 import { TeardownOutputDialog } from '@/components/worktree/TeardownOutputDialog'
 import { WindowResizeHandles } from './WindowResizeHandles'
 
@@ -801,6 +802,7 @@ export function MainWindow() {
         </Suspense>
       )}
       <BranchConflictDialog />
+      <ClearContextConfirmDialog />
       <TeardownOutputDialog />
       <Toaster
         position="bottom-right"